Simplify dbtest macros
