+++ sky-macros/src/lib.rs
@@ -43,255 +43,13 @@
 //!
 
 use proc_macro::TokenStream;
-use proc_macro2::Span;
-use quote::quote;
-use syn::{self};
 
-const CHARSET: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-
-/// This parses a function within a `dbtest` module
-///
-/// This accepts an `async` function and returns a non-`async` version of it - by
-/// making the body of the function use the `tokio` runtime
-fn parse_dbtest(
-    mut input: syn::ItemFn,
-    rng: &mut impl rand::Rng,
-    table_decl: Option<String>,
-) -> Result<TokenStream, syn::Error> {
-    let sig = &mut input.sig;
-    let fname = sig.ident.to_string();
-    let body = &input.block;
-    let attrs = &input.attrs;
-    let vis = &input.vis;
-    let header = quote! {
-        #[::core::prelude::v1::test]
-    };
-    if sig.asyncness.is_none() {
-        let msg = "`dbtest` functions need to be async";
-        return Err(syn::Error::new_spanned(sig.fn_token, msg));
-    }
-    sig.asyncness = None;
-    let rand_string: String = (0..64)
-        .map(|_| {
-            let idx = rng.gen_range(0..CHARSET.len());
-            CHARSET[idx] as char
-        })
-        .collect();
-    let table_decl = table_decl.unwrap_or_else(|| "keymap(str,str)".to_owned());
-    let body = quote! {
-        let mut con = skytable::AsyncConnection::new("127.0.0.1", 2003).await.unwrap();
-        let __create_ks =
-            con.run_simple_query(
-                &skytable::query!("create", "keyspace", "testsuite")
-            ).await.unwrap();
-        if !(
-            __create_ks == skytable::Element::RespCode(skytable::RespCode::Okay) ||
-            __create_ks == skytable::Element::RespCode(
-                skytable::RespCode::ErrorString(
-                    skytable::error::errorstring::ERR_ALREADY_EXISTS.to_owned()
-                )
-            )
-        ) {
-            panic!("Failed to create keyspace: {:?}", __create_ks);
-        }
-        let __switch_ks =
-            con.run_simple_query(
-                &skytable::query!("use", "testsuite")
-            ).await.unwrap();
-        if (__switch_ks != skytable::Element::RespCode(skytable::RespCode::Okay)) {
-            panic!("Failed to switch keyspace: {:?}", __switch_ks);
-        }
-        let __create_tbl =
-            con.run_simple_query(
-                &skytable::query!(
-                    "create",
-                    "table",
-                    #rand_string,
-                    #table_decl,
-                    "volatile"
-                )
-            ).await.unwrap();
-        assert_eq!(
-            __create_tbl, skytable::Element::RespCode(skytable::RespCode::Okay), "Failed to create table"
-        );
-        let mut __concat_entity = std::string::String::new();
-        __concat_entity.push_str("testsuite:");
-        __concat_entity.push_str(&#rand_string);
-        let __MYENTITY__: String = __concat_entity.clone();
-        let __switch_entity =
-            con.run_simple_query(
-                &skytable::query!("use", __concat_entity)
-            ).await.unwrap();
-        assert_eq!(
-            __switch_entity, skytable::Element::RespCode(skytable::RespCode::Okay), "Failed to switch"
-        );
-        let mut query = skytable::Query::new();
-        #body
-        {
-            let mut __flush__ = skytable::Query::from("flushdb");
-            std::assert_eq!(
-                con.run_simple_query(&__flush__).await.unwrap(),
-                skytable::Element::RespCode(skytable::RespCode::Okay)
-            );
-        }
-    };
-    let result = quote! {
-        #header
-        #(#attrs)*
-        #vis #sig {
-            tokio::runtime::Builder::new_multi_thread()
-            .worker_threads(4)
-            .thread_name(#fname)
-            .thread_stack_size(3 * 1024 * 1024)
-            .enable_all()
-            .build()
-            .unwrap()
-            .block_on(async { #body });
-        }
-    };
-    Ok(result.into())
-}
-
-/// This function checks if the current function is eligible to be a test
-fn parse_test_sig(
-    input: syn::ItemFn,
-    rng: &mut impl rand::Rng,
-    table_decl: Option<String>,
-) -> TokenStream {
-    for attr in &input.attrs {
-        if attr.path.is_ident("test") {
-            let msg = "second test attribute is supplied";
-            return syn::Error::new_spanned(&attr, msg)
-                .to_compile_error()
-                .into();
-        }
-    }
-
-    if !input.sig.inputs.is_empty() {
-        let msg = "the test function cannot accept arguments";
-        return syn::Error::new_spanned(&input.sig.inputs, msg)
-            .to_compile_error()
-            .into();
-    }
-    parse_dbtest(input, rng, table_decl).unwrap_or_else(|e| e.to_compile_error().into())
-}
-
-/// This function accepts an entire module which comprises of `dbtest` functions.
-/// It takes each function in turn, and generates `#[test]`able functions for them
-fn parse_test_module(args: TokenStream, item: TokenStream) -> TokenStream {
-    let input = syn::parse_macro_input!(item as syn::ItemMod);
-    let content = match input.content {
-        Some((_, c)) => c,
-        None => {
-            return syn::Error::new_spanned(&input, "Couldn't get the module content")
-                .to_compile_error()
-                .into()
-        }
-    };
-    let args = syn::parse_macro_input!(args as syn::AttributeArgs);
-    let mut skips = Vec::new();
-    let mut table_decl = None;
-    for arg in args {
-        if let syn::NestedMeta::Meta(syn::Meta::NameValue(namevalue)) = arg {
-            let ident = namevalue.path.get_ident();
-            if ident.is_none() {
-                let msg = "Must have specified ident";
-                return syn::Error::new_spanned(namevalue, msg)
-                    .to_compile_error()
-                    .into();
-            }
-            match ident.unwrap().to_string().to_lowercase().as_str() {
-                "skip" => {
-                    let skip_lit = namevalue.lit.clone();
-                    let span = skip_lit.span();
-                    skips = match parse_string(skip_lit, span, "skip") {
-                        Ok(s) => s,
-                        Err(_) => {
-                            return syn::Error::new_spanned(
-                                namevalue,
-                                "Expected a value for argument `skip`",
-                            )
-                            .to_compile_error()
-                            .into();
-                        }
-                    }
-                    .split_whitespace()
-                    .map(|val| val.to_string())
-                    .collect();
-                }
-                "table" => {
-                    // check if the user wants some special table declaration
-                    let table_decl_lit = namevalue.lit.clone();
-                    let span = table_decl_lit.span();
-                    table_decl = match parse_string(table_decl_lit, span, "table") {
-                        Ok(s) => Some(s),
-                        Err(_) => {
-                            return syn::Error::new_spanned(
-                                namevalue,
-                                "Expected a value for argument `table`",
-                            )
-                            .to_compile_error()
-                            .into();
-                        }
-                    }
-                }
-                x => {
-                    let msg = format!("Unknown attribute {} is specified; expected `skip`", x);
-                    return syn::Error::new_spanned(namevalue, msg)
-                        .to_compile_error()
-                        .into();
-                }
-            }
-        }
-    }
-    let mut result = quote! {};
-    let mut rng = rand::thread_rng();
-    for item in content {
-        match item {
-            // We just care about functions, so parse functions and ignore everything
-            // else
-            syn::Item::Fn(function) => {
-                if skips.contains(&function.sig.ident.to_string()) {
-                    result = quote! {
-                        #result
-                        #function
-                    };
-                    continue;
-                }
-                let inp = parse_test_sig(function, &mut rng, table_decl.clone());
-                let __tok: syn::ItemFn = syn::parse_macro_input!(inp as syn::ItemFn);
-                let tok = quote! {
-                    #__tok
-                };
-                result = quote! {
-                    #result
-                    #tok
-                };
-            }
-            token => {
-                result = quote! {
-                    #result
-                    #token
-                };
-            }
-        }
-    }
-    result.into()
-}
-
-fn parse_string(int: syn::Lit, span: Span, field: &str) -> Result<String, syn::Error> {
-    match int {
-        syn::Lit::Str(s) => Ok(s.value()),
-        syn::Lit::Verbatim(s) => Ok(s.to_string()),
-        _ => Err(syn::Error::new(
-            span,
-            format!("Failed to parse {} into a string.", field),
-        )),
-    }
-}
+mod dbtest_fn;
+mod dbtest_mod;
+mod util;
 
 #[proc_macro_attribute]
-/// The `dbtest` macro starts an async server in the background and is meant for
+/// The `dbtest_mdule` macro starts an async server in the background and is meant for
 /// use within the `skyd` or `WORKSPACEROOT/server/` crate. If you use this compiler
 /// macro in any other crate, you'll simply get compilation errors
 ///
@@ -317,6 +75,6 @@ fn parse_string(int: syn::Lit, span: Span, field: &str) -> Result<String, syn::E
 /// are practially impossible. Hence we do not bother with a global random string table and instead proceed
 /// to generate tables randomly at the point of invocation
 ///
-pub fn dbtest(args: TokenStream, item: TokenStream) -> TokenStream {
-    parse_test_module(args, item)
+pub fn dbtest_module(args: TokenStream, item: TokenStream) -> TokenStream {
+    dbtest_mod::parse_test_module(args, item)
 }
